requirements.txt
langchain
pdfplumber
torch
transformers
weaviate-client
python-dotenv